Add new notebook for data exploration in research section
This commit introduces a new Jupyter notebook to the research-and-development section. The notebook includes code to import necessary modules, read in CSV data using pandas, display dataset info, select features that are not objects, and print feature names. It also provides lines of code for further data exploration which are commented out.